workink version, need to clean leacks

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: apenrose <apenrose@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/29 19:39:24 by apenrose          #+#    #+#             */
-/*   Updated: 2020/12/09 00:50:50 by apenrose         ###   ########.fr       */
+/*   Updated: 2020/12/10 23:53:18 by apenrose         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len);
 int		ft_strchr(const char *s, int c);
 char	*ft_strdup(const char *s1);
 size_t	ft_strlenn(const char *s);
-size_t	ft_strlen(const char *s);
+int	    ft_strlen(const char *s);
 char	*part_two(char *s1, char *s2, char *uni);
 char	*ft_strjoin(char *s1, char *s2);
